requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
jedi==0.17.2
numpy==1.21.1
pandas==1.3.1
parso==0.7.1
pluggy==1.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
ujson==4.1.0
urllib3==1.26.6
